refactor: extract AddLeatherColorForm, clean up threadColorOperations.server.js logs, and modularize add leather color route
